git backed session file
use git to keep history of session file instead of copying file to
session-old. This is only run on program start and is not a replacment
for autosave. keeping old version of session backup behind #ifdef, may
be removed later or switched to a runtime flag.

issues, when using a new session file and no .git exist, first run will
have a pathspec error from git. not an issue as git is just trying to
checkout a nonexistent file and is fixed once git commits the session
file on the next startup.