Remove desktop integration
personally not interested in this code and even when disabled there is
still desktop code that is build. it is interesting and have played with
it but if you want a file manager to provide a desktop then Microsoft
provides explorer.exe

could still use logic cleanup where there were if(browser) else(desktop)
checkes. with only the if(broswer) check left it may no longer be needed.